Pass the compilers using CMAKE_XXX_COMPILER instead of XXX env variable
The previous method of setting the CC and friends environment variables causes some issues
with some vcpkg ports. Sometimes a port must be compiled for the host, but when the CC env
variable is set, this overwrites all settings. This leads to a situation where code that
must be compiled for the host architecture (eg amd64) will be compiled with an arm64
cross-compiler.