fix: utilize vcpkg release target triplets, set host triplets
Switch to vcpkg release triplets to prevent building debug versions of dependencies, which are unused in DuckDB extensions. This reduces build times by avoiding unnecessary debug builds.

Specify the vcpkg host triplet in the build matrix to improve binary caching. vcpkg may rebuild dependencies when host and target triplets differ, so aligning them with the release triplet helps prevent this. The exception is the ARM64 Linux target built on an x86 Linux host, but this will change once GitHub ARM runners are available.