Add new triplet for Windows compatible with VS2019
This change adds a new VCPKG triplet
`x64-windows-static-md-release-vs2019comp` (and a toolchain definition
for it) that is intended to be used for all MSVC Windows builds
(currently only `x64-windows-static-md-release`).

This custom triplet appeared to bes necessary to pass
`/D_DISABLE_CONSTEXPR_MUTEX_CONSTRUCTOR` flag
to AWS C++ SDK compilation. This flag enables the compatibility with
older C++ stdlib from VS2019. It was added to the `duckdb/duckdb` repo
in #17991. That addition effectively propagated this flag to the
extensions sources, but not to extension dependencies built with VCPKG.
The problem with `duckdb-aws` was uncovered while running tests added
in #18149 PR.