Add dependency between build and metadata targets
It seems that with C API extension setup, the
`build_extension_with_metadata_debug` target does not have a dependency
on the target that actually builds the library. So if the `make` is for
some reason invoked in parallel mode with `-j` option then the metadata
target fails like this:

```python
Traceback (most recent call last):
  File "/path/to/extension-ci-tools/scripts/append_extension_metadata.py", line 121, in <module>
    main()
    ~~~~^^
  File "/path/to/extension-ci-tools/scripts/append_extension_metadata.py", line 57, in main
    shutil.copyfile(args.library_file, OUTPUT_FILE_TMP)
    ~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/lib64/python3.13/shutil.py", line 260, in copyfile
    with open(src, 'rb') as fsrc:
         ~~~~^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: './build/release/libmy_ext.so'
```

Perhaps this is not a realistic scenario, because, to actually enable
parallel builds the top-level `make` should to be invoked without `-j`
like one of the following variants:

```
GEN=ninja make
```
```
CMAKE_BUILD_PARALLEL_LEVEL=<num_cpus> make
```

Still, assuming the `build_extension_library_debug` target exists for
all toolchains (currently, C/C++ and Rust) that may use the
`build_extension_with_metadata_debug` then it may be better to add an
explicit dependency on it.

The same is also applied to metadata and build `_release` targets.