Enable back Linux test runs
I was chasing a rabbit hole about extension CI test runs from [this comment](https://github.com/duckdb/extension-template-rs/pull/35#issuecomment-3588918205)
and noticed that test runs are disabled for a few Linux configurations.

I think it is possible that the reasons for disabling them there are
outdated, so filing a PR to see whether these tests can be run now.